Cleanup Columbia First Push
This app was created using Esri’s Runtime SDK for QT library. The purpose of this app was to give the Cleanup Columbia crew managers a way to easily check in groups of the cleanup crews into certain cleanup locations. The reason this offline, native app was necessary was because several locations for Cleanup Columbia would not have online capabilities. This app was intended to make offline edits to the Cleanup Columbia services, and push the edits online after a network connection is re-established